Fix PG metric type bug + migration v11 format string bug
- Fixed 'invalid input syntax for type integer: 0.0' in metric recording
  (round() returns float, cast to int() before passing to increment_metric)
- Fixed migration v11 crash: 'DEFAULT '{}'' Python format string conflict
  (double-escaped to '{{}}' so .format() doesn't interpret as placeholder)
- Migration v10 pgvector error is harmless (optional extension)

+++ src/memory_bridge/main.py
@@ -256,7 +256,7 @@ def create_app() -> FastAPI:
         try:
             storage = await get_storage()
             await storage.increment_metric("request_count")
-            await storage.increment_metric("total_latency_ms", round(elapsed_ms, 1))
+            await storage.increment_metric("total_latency_ms", int(round(elapsed_ms, 1)))
         except Exception:
             logger.exception("Failed to record storage metrics")
